% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{myprime}
\alias{myprime}
\title{Checking prime number}
\usage{
myprime(x)
}
\description{
The name of the function is myprime.
The format of the function is myprime(x).
This function checks prime numbers.
In other words, If you put one of the counting numbers to x, it tells you whether the number is a prime number or not.
}
\examples{

myprime(11)
}
